+++ frontend/components/traces/sessions-table/columns.tsx
@@ -4,8 +4,8 @@ import { ChevronDownIcon, ChevronRightIcon } from "lucide-react";
 import ClientTimestampFormatter from "@/components/client-timestamp-formatter";
 import { ColumnFilter } from "@/components/ui/datatable-filter/utils";
 import Mono from "@/components/ui/mono";
-import { getDurationString } from "@/lib/flow/utils";
 import { SessionPreview, Trace } from "@/lib/traces/types";
+import { getDurationString } from "@/lib/utils";
 
 type SessionRow = {
   type: string;
